+++ js/main.js
@@ -16,10 +16,10 @@ import './filter.js';
 import {setUserFormSubmit} from './api.js';
 import {openSuccessModule} from './success-modal.js';
 import {openErrorModule} from './error-modal.js';
-import {getData} from './api.js';
+// import {getData} from './api.js';
 import {adForm} from './form-validation.js';
 
-getData();
+// getData();
 
 adForm.addEventListener('submit', (evt) => {
   evt.preventDefault();
