+++ js/utils/scroll.js
@@ -10,8 +10,7 @@ const unblockScroll = () => {
   const scrollY = body.style.top;
   body.style.position = '';
   body.style.top = '';
-  // eslint-disable-next-line radix
-  window.scrollTo(0, parseInt(scrollY || '0') * -1);
+  window.scrollTo(0, parseInt(scrollY || '0', 10) * -1);
 };
 
 export {blockScroll};
